Add income_from_player to score changes
This commit adds the `income_from_player` field to the `PlayerScoreChange` model and the
`ScoreChangeEvent` API model. It also updates the `change_player_score` function to accept an
optional `income_from_player` argument, which is used to record the player who provided the
income. This information is then included in the `ScoreChangeEvent` that is returned by the
API.